Stop get_columns returning nils when column_names are UUIDs
When an array of UUIDs is passed as the column_names argument to get_columns (i.e.  because the column family's CompareWith is a UUID type), an array of nils is returned. This is because the values in column_names are munged to their string representations, yet the Thrift binding returns a hash where the keys are actual UUID objects:

(rdb:1) where
--> #0 Cassandra::Protocol._get_columns(column_family#String, key#String, columns#Array,...)
       at line /Users/aanand/.rvm/gems/ruby-1.9.1-p378/gems/cassandra-0.7.5/lib/cassandra/protocol.rb:41
    #1 Cassandra.get_columns(column_family#String, key#String)
       at line /Users/aanand/.rvm/gems/ruby-1.9.1-p378/gems/cassandra-0.7.5/lib/cassandra/cassandra.rb:170
    #2 CassandraTest.test_get_multiple_time_uuid_columns
       at line test/cassandra_test.rb:62
(rdb:1) list
[36, 45] in /Users/aanand/.rvm/gems/ruby-1.9.1-p378/gems/cassandra-0.7.5/lib/cassandra/protocol.rb
   36          columns_to_hash(column_family, @client.get_slice(@keyspace, key,
   37            CassandraThrift::ColumnParent.new(:column_family => column_family),
   38            CassandraThrift::SlicePredicate.new(:column_names => columns),
   39            consistency))
   40        end
=> 41        (sub_columns || columns).map { |name| result[name] }
   42      end
   43
   44      def _multiget(column_family, keys, column, sub_column, count, start, finish, reversed, consistency)
   45        # Single values; count and range parameters have no effect
(rdb:1) p columns
["\x93\x81@\x00\xB6h\x11\xB2\x9A\x1FW\th\xB0^o", "\x13\x81@\x00N\xFF\x11\xB3\x96\xEF0\xA9\xE0\xA3:F"]
(rdb:1) p result.keys
[<UUID#10152730 time: 1970-07-14 00:20:16 -0400, usecs: 0 jitter: 8025045546900594586>, <UUID#10152632 time: 1971-01-24 03:40:32 -0500, usecs: 0 jitter: 5060537316610731926>]

This is fixed by converting column_names' values back to UUIDs (or whatever) before doing the hash lookup.

I expected to have to do the same for sub_columns, but oddly enough the Thrift binding returns a string-keyed hash in that case, so everything still works. It's not clear to me whether or not that's correct behaviour.