Changed composite slice args to [:before, :after].
Inclusive and exclusive didn't really work b/c the values
are different if it is the start of the slice vs the end
of a slice. Before/after better represent what this actually does,
appending either a negative or positive value between composite
components.

Fixed comparison logic so that the mock class works.